Straighten out the datatypes
You can't have a variable-length struct, so I'd have to specify whether an instruction could possibly use extra clock cycles.
When you consider that we have to return a flag from the execute function to signal whether those extra cycles are used, it makes far more sense to simply return the number of extra cycles.
Usually that number is just 0, but sometimes it can be 6. Moving that flag from the struct and into the return value fixes the compile errors and should also save a bit of memory.