Overengineer the run script
It may be overengineered, but I think it's worthwhile. The script will now also take care of mounting the Pico if you haven't already done so.
This makes it very quick to upload code, which is great for development.
Pretty much as soon as you plug in the Pico, you can hit enter and it will take care of the rest for you! Very handy indeed.